Add backward support for " ./scripts/bpf_helpers_doc.py"

+++ scripts/package_building/build_artifacts.sh
@@ -490,7 +490,15 @@ function prepare_perf_rhel (){
         cp -fr ./tools "${base_dir}/perf/rpmbuild/SOURCES"
         cp -fr ./include "${base_dir}/perf/rpmbuild/SOURCES"
         mkdir "${base_dir}/perf/rpmbuild/SOURCES/scripts"
-        cp -fr ./scripts/bpf_doc.py "${base_dir}/perf/rpmbuild/SOURCES/scripts"
+	if [ -e ./scripts/bpf_doc.py ]
+	then
+        	cp -fr ./scripts/bpf_doc.py "${base_dir}/perf/rpmbuild/SOURCES/scripts"	
+	fi
+	if [ -e  ./scripts/bpf_helpers_doc.py ]
+	then
+        	cp -fr  ./scripts/bpf_helpers_doc.py "${base_dir}/perf/rpmbuild/SOURCES/scripts"	
+	fi
+	
     fi
     popd
     pushd "${base_dir}/perf/rpmbuild"
